Add procedural terrain sync contracts
Expose procedural terrain profile and sync manifest APIs for fun-ecs.

Add deterministic page digest probes and validation errors for client/server procedural page checks.

Keep the widened terrain generator version type compiling across source, terrain, and API surfaces.

+++ fun-ecs/src/spatial/source.rs
@@ -142,7 +142,7 @@ pub struct EcsProceduralRecipeRef {
     pub key: EcsSpatialPageKey,
     pub recipe_id: u64,
     pub seed: u64,
-    pub generator_version: u16,
+    pub generator_version: u32,
     pub manifest_signature: u64,
     pub checksum: EcsSourceChecksum,
 }
